Move volume/local to components/volume/local
